backend/requirements.txt
factory_boy
djangorestframework
django-cors-headers
